bond: fix type for port and queue id
In some places port_id is declared as int or uint16_t, which causes data
loss when using functions expecting port_id as uint8_t.

Previous type of rx and tx queue_id field was int, which caused possible
data loss when calling functions expecting uint16_t as an argument.

Issues found with static code analysis tool.

+++ lib/librte_pmd_bond/rte_eth_bond_8023ad.c
@@ -747,7 +747,7 @@ bond_mode_8023ad_periodic_cb(void *arg)
 	struct ether_addr slave_addr;
 
 	void *pkt = NULL;
-	uint16_t i, slave_id;
+	uint8_t i, slave_id;
 
 
 	/* Update link status on each port */
@@ -1050,7 +1050,7 @@ int
 bond_mode_8023ad_enable(struct rte_eth_dev *bond_dev)
 {
 	struct bond_dev_private *internals = bond_dev->data->dev_private;
-	uint16_t i;
+	uint8_t i;
 
 	for (i = 0; i < internals->active_slave_count; i++)
 		bond_mode_8023ad_activate_slave(bond_dev, i);
